feat: clean build cache
