+++ apps/ext/package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "start": "WEB_PORT=3180 webpack serve",
     "clean": "yarn clean:build",
-    "clean:build": "rimraf ./build && rimraf .tamagui",
+    "clean:build": "rimraf ./build && rimraf .tamagui && rimraf ./node_modules/.cache",
     "build": "cross-env NODE_ENV=production webpack build",
     "build:all": "yarn clean && yarn build && node ../../development/webpack/ext/zip.js",
     "start:v3": "EXT_MANIFEST_V3=1 WEB_PORT=3180 webpack serve",
