+++ apps/mobile/package.json
@@ -11,9 +11,8 @@
     "ios": "expo run:ios",
     "ios:device": "expo run:ios --device",
     "native-bundle": "react-native start",
-    "clean": " yarn clean:build",
-    "clean:build": "rimraf ./dist && rimraf ./expo && rimraf .tamagui",
-    "clean:cache": "rm -rf $TMPDIR/metro-bundler-cache-* && rm -rf $TMPDIR/metro-file-map-* && rm -rf $TMPDIR/metro-cache ",
+    "clean": "yarn clean:build",
+    "clean:build": "rimraf ./dist && rimraf ./expo && rimraf .tamagui && rimraf ./node_modules/.cache",
     "split-bundle": "SPLIT_BUNDLE=1 react-native start",
     "native-bundle:analyze": "react-native-bundle-visualizer"
   },
