+++ apps/web-embed/package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "start": "webpack serve",
     "clean": "yarn clean:build",
-    "clean:build": "rimraf ./web-build && rimraf .tamagui",
+    "clean:build": "rimraf ./web-build && rimraf .tamagui && rimraf ./node_modules/.cache",
     "build": "cross-env NODE_ENV=production webpack build && cp ./web-build/index.html ./web-build/404.html && sh ./postbuild.sh"
   },
   "private": true,
